make getting all players for user with ftlh rename PlayerController to PlayerRestController
