+++ registration/src/main/java/com/team1/registration/controllers/player/PlayerController.java
@@ -1,62 +1,31 @@
 package com.team1.registration.controllers.player;
 
-import com.team1.registration.models.Player;
-import com.team1.registration.models.Team;
+
 import com.team1.registration.services.PlayerService;
-import com.team1.registration.services.TeamService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.ModelAndView;
 
-import java.util.List;
-import java.util.Objects;
-
-@RestController
+@Controller
+@Slf4j
 @AllArgsConstructor
 @RequestMapping("/players")
-@Slf4j
 public class PlayerController {
     private PlayerService playerService;
-    private TeamService teamService;
 
-    @PostMapping
-    @ResponseStatus(value = HttpStatus.OK, reason = "player created")
-    public void registerPlayer(@RequestBody Player player) {
-        playerService.registerPlayer(player);
-    }
+    //todo: move createPlayer from ProfileService here
 
-    @GetMapping
-    public List<Player> getAllPlayers() {
-        return playerService.getAllPlayers();
-    }
-
-    @GetMapping("/{playerId}")
-    public Player getPlayerById(@PathVariable Integer playerId) {
-        return playerService.getPlayerById(playerId);
-    }
-
-    // todo: change endpoint
-    @ResponseStatus(value = HttpStatus.NO_CONTENT, reason = "player joined to team")
-    @PostMapping("/{playerId}/join-team/{teamId}")
-    public void joinTeam(@PathVariable Integer playerId, @PathVariable Integer teamId) {
-        var player = playerService.getPlayerById(playerId);
-        var team = teamService.getTeamById(teamId);
-        teamService.addPlayerToTeam(team, player);
-        teamService.updateTeam(team);
-        playerService.addTeamToPlayer(player, team);
-        playerService.updatePlayer(player);
-    }
-
-    @ResponseStatus(value = HttpStatus.NO_CONTENT, reason = "new team created")
-    @PostMapping("{playerId}/new-team")
-    public void createTeam(@PathVariable Integer playerId, @RequestBody Team team) {
-        if (!playerService.containsPlayer(playerId)) {
-            return;
-        }
-        //todo: add response status for the case of not finding the player
-        team.setCreatorId(playerId);
-        teamService.registerTeam(team);
-        this.joinTeam(playerId, team.getId());
+    //fixme: do it without @PathVariable
+    @GetMapping("{userId}")
+    public ModelAndView getAllUserPlayers(@PathVariable Integer userId) {
+        var players = playerService.getPlayersByUserId(userId);
+        var modelAndView = new ModelAndView();
+        modelAndView.addObject("players", players);
+        modelAndView.setViewName("players/getAllPlayers");
+        return modelAndView;
     }
 }
