+++ registration/src/main/java/com/team1/registration/controllers/team/TeamController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/teams")
+@RequestMapping("/teams-rest")
 @AllArgsConstructor
 public class TeamController {
     private TeamService teamService;
