+++ registration/src/main/resources/templates/players/getAllPlayers.ftlh
@@ -1,0 +1,8 @@
+<#import "../parts/common.ftlh" as c>
+<#include "../parts/security.ftlh">
+
+<@c.page>
+    <#list players as p>
+        ${p} <br>
+    </#list>
+</@c.page>
